ssdfs: Fix bugs in ZNS SSD support implementation (step 5)
Fix bugs in ZNS SSD support implementation (step 5).

+++ fs/ssdfs/peb_block_bitmap.h
@@ -121,10 +121,10 @@ int ssdfs_peb_blk_bmap_get_invalid_pages(struct ssdfs_peb_blk_bmap *ptr);
 int ssdfs_peb_define_reserved_pages_per_log(struct ssdfs_peb_blk_bmap *bmap);
 int ssdfs_peb_blk_bmap_reserve_metapages(struct ssdfs_peb_blk_bmap *bmap,
 					 int bmap_index,
-					 u16 count);
+					 u32 count);
 int ssdfs_peb_blk_bmap_free_metapages(struct ssdfs_peb_blk_bmap *bmap,
 				      int bmap_index,
-				      u16 count);
+				      u32 count);
 int ssdfs_peb_blk_bmap_pre_allocate(struct ssdfs_peb_blk_bmap *bmap,
 				    int bmap_index,
 				    u32 *len,
